// backend/src/rules_engine/domain/enums/currency.enum.ts
export enum Currency {
  USD = 'USD',
  EUR = 'EUR',
  GBP = 'GBP',
  JPY = 'JPY',
  AUD = 'AUD',
  CAD = 'CAD',
  CHF = 'CHF',
  CNY = 'CNY',
  SEK = 'SEK',
  NZD = 'NZD',
  CLP = 'CLP',
  MXN = 'MXN',
  ARS = 'ARS',
  BRL = 'BRL',
  COP = 'COP',
  PEN = 'PEN',
  VEF = 'VEF',
  UYU = 'UYU',
  PYG = 'PYG',
  DOP = 'DOP',
  CRC = 'CRC',
  GTQ = 'GTQ',
  HNL = 'HNL',
  NIO = 'NIO',
  RON = 'RON',
  RUB = 'RUB',
  TRY = 'TRY',
  ZAR = 'ZAR',
}